gameを作成できる

+++ db/migrate/20240104124847_create_games.rb
@@ -9,7 +9,7 @@ class CreateGames < ActiveRecord::Migration[7.1]
       t.string :black_id
       t.string :black_rating
       t.integer :black_k
-      t.integer :white_point
+      t.float :white_point
       t.references :tournament, foreign_key: true
 
       t.timestamps
